refactor: extract operation options to options.go and add input validation
- Rename reboot.go to civo.go to better reflect the Civo executor scope
- Extract FOP (Option, WithAPIConfig, WithClient) into options.go
- Add validation for clusterID, apiKey, and apiURL before Civo client creation
