+++ pkg/operation/civo.go
@@ -8,27 +8,6 @@ import (
 	"github.com/civo/civogo"
 )
 
-// Option represents a configuration function that modifies civoExecutor.
-type Option func(*civoExecutor)
-
-// WithAPIConfig returns Option to configure the Civo API credentials and version.
-// The client is created internally using these values.
-func WithAPIConfig(apiKey, apiURL, region, version string) Option {
-	return func(e *civoExecutor) {
-		e.apiKey = apiKey
-		e.apiURL = apiURL
-		e.region = region
-		e.version = version
-	}
-}
-
-// WithClient returns Option to inject a pre-built Civo client (for testing).
-func WithClient(client civogo.Clienter) Option {
-	return func(e *civoExecutor) {
-		e.civoClient = client
-	}
-}
-
 // civoExecutor implements Executor using the Civo API.
 type civoExecutor struct {
 	civoClient civogo.Clienter
@@ -47,8 +26,19 @@ func NewCivoExecutor(clusterID string, opts ...Option) (Executor, error) {
 		opt(e)
 	}
 
+	if clusterID == "" {
+		return nil, fmt.Errorf("cluster ID must not be empty")
+	}
+
 	if e.civoClient != nil {
 		return e, nil
+	}
+
+	if e.apiKey == "" {
+		return nil, fmt.Errorf("API key must not be empty")
+	}
+	if e.apiURL == "" {
+		return nil, fmt.Errorf("API URL must not be empty")
 	}
 
 	client, err := civogo.NewClientWithURL(e.apiKey, e.apiURL, e.region)
